WEEK005-学习JHipster/demo/src/test/javascript/spec/app/entities/foo/foo-update.component.spec.ts
/* tslint:disable max-line-length */
import { shallowMount, createLocalVue, Wrapper } from '@vue/test-utils';
import sinon, { SinonStubbedInstance } from 'sinon';
import Router from 'vue-router';
import { ToastPlugin } from 'bootstrap-vue';

import * as config from '@/shared/config/config';
import FooUpdateComponent from '@/entities/foo/foo-update.vue';
import FooClass from '@/entities/foo/foo-update.component';
import FooService from '@/entities/foo/foo.service';

import AlertService from '@/shared/alert/alert.service';

const localVue = createLocalVue();

config.initVueApp(localVue);
const i18n = config.initI18N(localVue);
const store = config.initVueXStore(localVue);
const router = new Router();
localVue.use(Router);
localVue.use(ToastPlugin);
localVue.component('font-awesome-icon', {});
localVue.component('b-input-group', {});
localVue.component('b-input-group-prepend', {});
localVue.component('b-form-datepicker', {});
localVue.component('b-form-input', {});

describe('Component Tests', () => {
  describe('Foo Management Update Component', () => {
    let wrapper: Wrapper<FooClass>;
    let comp: FooClass;
    let fooServiceStub: SinonStubbedInstance<FooService>;

    beforeEach(() => {
      fooServiceStub = sinon.createStubInstance<FooService>(FooService);

      wrapper = shallowMount<FooClass>(FooUpdateComponent, {
        store,
        i18n,
        localVue,
        router,
        provide: {
          fooService: () => fooServiceStub,
          alertService: () => new AlertService(),
        },
      });
      comp = wrapper.vm;
    });

    describe('save', () => {
      it('Should call update service on save for existing entity', async () => {
        // GIVEN
        const entity = { id: 123 };
        comp.foo = entity;
        fooServiceStub.update.resolves(entity);

        // WHEN
        comp.save();
        await comp.$nextTick();

        // THEN
        expect(fooServiceStub.update.calledWith(entity)).toBeTruthy();
        expect(comp.isSaving).toEqual(false);
      });

      it('Should call create service on save for new entity', async () => {
        // GIVEN
        const entity = {};
        comp.foo = entity;
        fooServiceStub.create.resolves(entity);

        // WHEN
        comp.save();
        await comp.$nextTick();

        // THEN
        expect(fooServiceStub.create.calledWith(entity)).toBeTruthy();
        expect(comp.isSaving).toEqual(false);
      });
    });

    describe('Before route enter', () => {
      it('Should retrieve data', async () => {
        // GIVEN
        const foundFoo = { id: 123 };
        fooServiceStub.find.resolves(foundFoo);
        fooServiceStub.retrieve.resolves([foundFoo]);

        // WHEN
        comp.beforeRouteEnter({ params: { fooId: 123 } }, null, cb => cb(comp));
        await comp.$nextTick();

        // THEN
        expect(comp.foo).toBe(foundFoo);
      });
    });

    describe('Previous state', () => {
      it('Should go previous state', async () => {
        comp.previousState();
        await comp.$nextTick();

        expect(comp.$router.currentRoute.fullPath).toContain('/');
      });
    });
  });
});
